+++ lib/generators/boring/graphql/install/templates/hello_world_resolver.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class Resolvers::HelloWorldResolver < Resolvers::BaseResolver
+  description "Returns hello world"
+
+  type String, null: false
+
+  def resolve
+    "Hello World"
+  end
+end